requirements.txt
pydantic[email]==1.8.2
fastapi==0.68.0
uvicorn==0.14.0
SQLAlchemy==1.4.26
pydantic==1.8.2
sqlalchemy-utils==0.37.8
python-dotenv==0.19.0
faker==8.11.0
fastapi-pagination==0.8.3
alembic==1.6.5
requests==2.26.0
pytest-asyncio
asgi_lifespan==1.0.1
asyncpg==0.24.0
elasticsearch[async]==7.14.0
aioredis[hiredis]==2.0.0
redis==3.5.3
gunicorn==20.1.0
httpx==0.19.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
fastapi-cache2[redis]==0.1.8
loguru
motor==2.3.0
mypy-extensions==0.4.3
pathspec==0.8.1
pycodestyle==2.6.0
pyflakes==2.2.0
pymongo==3.11.1